move request logic to reqeust.rs
